backend/requirements.txt
requirements.txt for the Django backend:

django>=6.0
djangorestframework>=3.17
django-cors-headers>=4.9
